requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-cuser==2017.3.16
django-filter==2.3.0
django-widget-tweaks==1.4.8
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
